+++ built/grammar/GrammarBackedCell.js
@@ -1,5 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+const GrammarConstants_1 = require("./GrammarConstants");
 /*
 A cell contains a word but also the type information for that word.
 */
@@ -41,40 +42,40 @@ class GrammarBackedCell {
         const context = fullLine.split(" ")[0]; // todo: XI
         if (word === undefined)
             return {
-                kind: "unfilledColumnError",
+                kind: GrammarConstants_1.default.errors.unfilledColumnError,
                 subkind: type,
                 level: index,
                 context: context,
-                message: `unfilledColumnError "${type}" column in "${fullLine}" at line ${line} column ${index}. Expected pattern: "${this._expectedLinePattern}". definition: ${this._node.getDefinition().toString()}`
+                message: `${GrammarConstants_1.default.errors.unfilledColumnError} "${type}" column in "${fullLine}" at line ${line} column ${index}. Expected pattern: "${this._expectedLinePattern}". definition: ${this._node.getDefinition().toString()}`
             };
         if (type === undefined)
             return {
-                kind: "extraWordError",
+                kind: GrammarConstants_1.default.errors.extraWordError,
                 subkind: fullLine,
                 level: index,
                 context: context,
-                message: `extraWordError "${word}" in "${fullLine}" at line ${line} column ${index}. Expected pattern: "${this._expectedLinePattern}".`
+                message: `${GrammarConstants_1.default.errors.extraWordError} "${word}" in "${fullLine}" at line ${line} column ${index}. Expected pattern: "${this._expectedLinePattern}".`
             };
         const grammarProgram = this._grammarProgram;
         const runTimeGrammarBackedProgram = this._node.getProgram();
         const wordTypeClass = this._getWordTypeClass();
         if (!wordTypeClass)
             return {
-                kind: "grammarDefinitionError",
+                kind: GrammarConstants_1.default.errors.grammarDefinitionError,
                 subkind: type,
                 level: index,
                 context: context,
-                message: `grammarDefinitionError No column type "${type}" in grammar "${grammarProgram.getExtensionName()}" found in "${fullLine}" on line ${line}. Expected pattern: "${this._expectedLinePattern}".`
+                message: `${GrammarConstants_1.default.errors.grammarDefinitionError} No column type "${type}" in grammar "${grammarProgram.getExtensionName()}" found in "${fullLine}" on line ${line}. Expected pattern: "${this._expectedLinePattern}".`
             };
         const isValid = wordTypeClass.isValid(this._word, runTimeGrammarBackedProgram);
         return isValid
             ? undefined
             : {
-                kind: "invalidWordError",
+                kind: GrammarConstants_1.default.errors.invalidWordError,
                 subkind: type,
                 level: index,
                 context: context,
-                message: `invalidWordError in "${fullLine}" at line ${line} column ${index}. "${word}" does not fit in "${type}" column. Expected pattern: "${this._expectedLinePattern}".`
+                message: `${GrammarConstants_1.default.errors.invalidWordError} in "${fullLine}" at line ${line} column ${index}. "${word}" does not fit in "${type}" column. Expected pattern: "${this._expectedLinePattern}".`
             };
     }
 }